+++ pkg/cloudprovider/suite_test.go
@@ -215,8 +215,8 @@ var _ = BeforeEach(func() {
 	ec2Cache.Flush()
 	kubernetesVersionCache.Flush()
 	instanceTypeCache.Flush()
-	subnetProvider.ResetCache()
-	securityGroupProvider.ResetCache()
+	subnetProvider.Reset()
+	securityGroupProvider.Reset()
 	cloudProvider.instanceProvider.launchTemplateProvider.kubeDNSIP = net.ParseIP("10.0.100.10")
 	launchTemplateProvider.kubeDNSIP = net.ParseIP("10.0.100.10")
 
@@ -418,7 +418,7 @@ var _ = Describe("Allocation", func() {
 	})
 	Context("Provider Backwards Compatibility", func() {
 		It("should launch a node using provider defaults", func() {
-			provisioner := test.Provisioner(coretest.ProvisionerOptions{
+			provisioner = test.Provisioner(coretest.ProvisionerOptions{
 				Provider: v1alpha1.AWS{
 					AMIFamily:             aws.String(v1alpha1.AMIFamilyAL2),
 					SubnetSelector:        map[string]string{"*": "*"},
@@ -447,7 +447,7 @@ var _ = Describe("Allocation", func() {
 			Expect(*launchTemplate.Version).To(Equal("$Latest"))
 		})
 		It("should discover security groups by ID", func() {
-			provisioner := test.Provisioner(coretest.ProvisionerOptions{
+			provisioner = test.Provisioner(coretest.ProvisionerOptions{
 				Provider: v1alpha1.AWS{
 					AMIFamily:             aws.String(v1alpha1.AMIFamilyAL2),
 					SubnetSelector:        map[string]string{"*": "*"},
@@ -468,7 +468,7 @@ var _ = Describe("Allocation", func() {
 			))
 		})
 		It("should discover subnets by ID", func() {
-			provisioner := test.Provisioner(coretest.ProvisionerOptions{
+			provisioner = test.Provisioner(coretest.ProvisionerOptions{
 				Provider: v1alpha1.AWS{
 					AMIFamily:             aws.String(v1alpha1.AMIFamilyAL2),
 					SubnetSelector:        map[string]string{"aws-ids": "subnet-test1"},
@@ -486,7 +486,7 @@ var _ = Describe("Allocation", func() {
 			Expect(fake.SubnetsFromFleetRequest(createFleetInput)).To(ConsistOf("subnet-test1"))
 		})
 		It("should use the instance profile on the Provisioner when specified", func() {
-			provisioner := test.Provisioner(coretest.ProvisionerOptions{
+			provisioner = test.Provisioner(coretest.ProvisionerOptions{
 				Provider: v1alpha1.AWS{
 					AMIFamily:             aws.String(v1alpha1.AMIFamilyAL2),
 					SubnetSelector:        map[string]string{"*": "*"},
@@ -504,6 +504,71 @@ var _ = Describe("Allocation", func() {
 			Expect(fakeEC2API.CalledWithCreateLaunchTemplateInput.Len()).To(Equal(1))
 			input := fakeEC2API.CalledWithCreateLaunchTemplateInput.Pop()
 			Expect(*input.LaunchTemplateData.IamInstanceProfile.Name).To(Equal("overridden-profile"))
+		})
+	})
+	Context("Subnet Compatibility", func() {
+		// Note when debugging these tests -
+		// hard coded fixture data (ex. what the aws api will return) is maintained in fake/ec2api.go
+		It("should default to the cluster's subnets", func() {
+			ExpectApplied(ctx, env.Client, provisioner, nodeTemplate)
+			pod := ExpectProvisioned(ctx, env.Client, cluster, recorder, provisioningController, prov, coretest.UnschedulablePod(
+				coretest.PodOptions{NodeSelector: map[string]string{v1.LabelArchStable: v1alpha5.ArchitectureAmd64}}))[0]
+			ExpectScheduled(ctx, env.Client, pod)
+			Expect(fakeEC2API.CreateFleetBehavior.CalledWithInput.Len()).To(Equal(1))
+			input := fakeEC2API.CreateFleetBehavior.CalledWithInput.Pop()
+			Expect(input.LaunchTemplateConfigs).To(HaveLen(1))
+
+			foundNonGPULT := false
+			for _, v := range input.LaunchTemplateConfigs {
+				for _, ov := range v.Overrides {
+					if *ov.InstanceType == "m5.large" {
+						foundNonGPULT = true
+						Expect(v.Overrides).To(ContainElements(
+							&ec2.FleetLaunchTemplateOverridesRequest{SubnetId: aws.String("subnet-test1"), InstanceType: aws.String("m5.large"), AvailabilityZone: aws.String("test-zone-1a")},
+							&ec2.FleetLaunchTemplateOverridesRequest{SubnetId: aws.String("subnet-test2"), InstanceType: aws.String("m5.large"), AvailabilityZone: aws.String("test-zone-1b")},
+							&ec2.FleetLaunchTemplateOverridesRequest{SubnetId: aws.String("subnet-test3"), InstanceType: aws.String("m5.large"), AvailabilityZone: aws.String("test-zone-1c")},
+						))
+					}
+				}
+			}
+			Expect(foundNonGPULT).To(BeTrue())
+		})
+		It("should launch instances into subnet with the most available IP addresses", func() {
+			fakeEC2API.DescribeSubnetsOutput.Set(&ec2.DescribeSubnetsOutput{Subnets: []*ec2.Subnet{
+				{SubnetId: aws.String("test-subnet-1"), AvailabilityZone: aws.String("test-zone-1a"), AvailableIpAddressCount: aws.Int64(10),
+					Tags: []*ec2.Tag{{Key: aws.String("Name"), Value: aws.String("test-subnet-1")}}},
+				{SubnetId: aws.String("test-subnet-2"), AvailabilityZone: aws.String("test-zone-1a"), AvailableIpAddressCount: aws.Int64(100),
+					Tags: []*ec2.Tag{{Key: aws.String("Name"), Value: aws.String("test-subnet-2")}}},
+			}})
+			ExpectApplied(ctx, env.Client, provisioner, nodeTemplate)
+			pod := ExpectProvisioned(ctx, env.Client, cluster, recorder, provisioningController, prov, coretest.UnschedulablePod(coretest.PodOptions{NodeSelector: map[string]string{v1.LabelTopologyZone: "test-zone-1a"}}))[0]
+			ExpectScheduled(ctx, env.Client, pod)
+			createFleetInput := fakeEC2API.CreateFleetBehavior.CalledWithInput.Pop()
+			Expect(fake.SubnetsFromFleetRequest(createFleetInput)).To(ConsistOf("test-subnet-2"))
+		})
+		It("should launch instances into subnets that are excluded by another provisioner", func() {
+			fakeEC2API.DescribeSubnetsOutput.Set(&ec2.DescribeSubnetsOutput{Subnets: []*ec2.Subnet{
+				{SubnetId: aws.String("test-subnet-1"), AvailabilityZone: aws.String("test-zone-1a"), AvailableIpAddressCount: aws.Int64(10),
+					Tags: []*ec2.Tag{{Key: aws.String("Name"), Value: aws.String("test-subnet-1")}}},
+				{SubnetId: aws.String("test-subnet-2"), AvailabilityZone: aws.String("test-zone-1b"), AvailableIpAddressCount: aws.Int64(100),
+					Tags: []*ec2.Tag{{Key: aws.String("Name"), Value: aws.String("test-subnet-2")}}},
+			}})
+			nodeTemplate.Spec.SubnetSelector = map[string]string{"Name": "test-subnet-1"}
+			ExpectApplied(ctx, env.Client, provisioner, nodeTemplate)
+			podSubnet1 := ExpectProvisioned(ctx, env.Client, cluster, recorder, provisioningController, prov, coretest.UnschedulablePod())[0]
+			ExpectScheduled(ctx, env.Client, podSubnet1)
+			createFleetInput := fakeEC2API.CreateFleetBehavior.CalledWithInput.Pop()
+			Expect(fake.SubnetsFromFleetRequest(createFleetInput)).To(ConsistOf("test-subnet-1"))
+
+			provisioner = test.Provisioner(coretest.ProvisionerOptions{Provider: &v1alpha1.AWS{
+				SubnetSelector:        map[string]string{"Name": "test-subnet-2"},
+				SecurityGroupSelector: map[string]string{"*": "*"},
+			}})
+			ExpectApplied(ctx, env.Client, provisioner)
+			podSubnet2 := ExpectProvisioned(ctx, env.Client, cluster, recorder, provisioningController, prov, coretest.UnschedulablePod(coretest.PodOptions{NodeSelector: map[string]string{v1alpha5.ProvisionerNameLabelKey: provisioner.Name}}))[0]
+			ExpectScheduled(ctx, env.Client, podSubnet2)
+			createFleetInput = fakeEC2API.CreateFleetBehavior.CalledWithInput.Pop()
+			Expect(fake.SubnetsFromFleetRequest(createFleetInput)).To(ConsistOf("test-subnet-2"))
 		})
 	})
 })
